Add prefix adapter, AudioSR

+++ utils/torch_utilities.py
@@ -120,7 +120,7 @@ def create_alignment_path(duration: torch.Tensor, mask: torch.Tensor):
     cum_duration = torch.cumsum(duration, 1)
 
     cum_duration_flat = cum_duration.view(b * t_x)
-    path = create_mask_from_length(cum_duration_flat, t_y).to(mask.dtype)
+    path = create_mask_from_length(cum_duration_flat, t_y).float()
     path = path.view(b, t_x, t_y)
     # take the diff on the `t_x` axis
     path = path - torch.nn.functional.pad(
